# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "default": [],
            "items": {},
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "FailRequest": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "FailRequest",
        "type": "object"
      },
...
      "PartialResultEnvelope": {
        "additionalProperties": false,
        "description": "Worker result: JSON only, never pickle (D9).",
        "properties": {
          "input_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Input Sha256",
            "type": "string"
          },
          "kernel": {
            "title": "Kernel",
            "type": "string"
          },
          "kernel_version": {
            "title": "Kernel Version",
            "type": "string"
          },
          "n_rows": {
            "minimum": 1.0,
            "title": "N Rows",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "payload_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Payload Sha256",
            "type": "string"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeInfo"
          },
          "timings": {
            "$ref": "#/components/schemas/ResultTimings"
          }
        },
        "required": [
          "kernel",
          "kernel_version",
          "input_sha256",
          "n_rows",
          "payload",
          "payload_sha256",
          "timings",
          "runtime"
        ],
        "title": "PartialResultEnvelope",
        "type": "object"
      },
...
      "ResultAck": {
        "properties": {
          "acceptance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance"
          },
          "assignment_id": {
            "title": "Assignment Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "assignment_id",
          "status"
        ],
        "title": "ResultAck",
        "type": "object"
      },
      "ResultTimings": {
        "properties": {
          "compute_ms": {
            "minimum": 0.0,
            "title": "Compute Ms",
            "type": "number"
          },
          "download_ms": {
            "minimum": 0.0,
            "title": "Download Ms",
            "type": "number"
          },
          "total_ms": {
            "minimum": 0.0,
            "title": "Total Ms",
            "type": "number"
          }
        },
        "required": [
          "download_ms",
          "compute_ms",
          "total_ms"
        ],
        "title": "ResultTimings",
        "type": "object"
      },
...
      "RuntimeInfo": {
        "properties": {
          "bundle": {
            "title": "Bundle",
            "type": "string"
          },
          "kind": {
            "enum": [
              "pyodide",
              "cpython"
            ],
            "title": "Kind",
            "type": "string"
          },
          "numpy": {
            "title": "Numpy",
            "type": "string"
          },
          "pyodide": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pyodide"
          },
          "python": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Python"
          }
        },
        "required": [
          "kind",
          "numpy",
          "bundle"
        ],
        "title": "RuntimeInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ProofNet API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/worker/assignments/{assignment_id}/fail": {
      "post": {
        "operationId": "fail_assignment_api_v1_worker_assignments__assignment_id__fail_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Fail Assignment",
        "tags": [
          "worker"
        ]
      }
    },
    "/api/v1/worker/assignments/{assignment_id}/input": {
      "get": {
        "operationId": "assignment_input_api_v1_worker_assignments__assignment_id__input_get",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Assignment Input",
        "tags": [
          "worker"
        ]
      }
    },
    "/api/v1/worker/assignments/{assignment_id}/result": {
      "post": {
        "operationId": "submit_result_api_v1_worker_assignments__assignment_id__result_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialResultEnvelope"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Submit Result",
        "tags": [
          "worker"
        ]
      }
    },
    "/api/v1/worker/assignments/{assignment_id}/start": {
      "post": {
        "operationId": "start_assignment_api_v1_worker_assignments__assignment_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Start Assignment",
        "tags": [
          "worker"
        ]
      }
    },
...
  }
}